Generate a table for FDA define.pdf document

Generate a table for FDA define.pdf document

fda_table(x, widths = c(0.75, 2.1, 0.6, 2.2), ...)

fda_table_file(file)

Arguments

x

a yspec object

widths

column widths in inches; must be numeric vector with length 4

...

not currently used

file

the full path to yaml specification file

Value

Character vector of latex code for the content of an FDA define.pdf document. It includes a table of contents as well as data spec tables for each dataset for a project.

Examples

spec <- load_spec_ex()
spec
#>  name    info unit          short                         source       
#>  C       c--  .             commented row indicator       _yspec_lookup
#>  USUBJID c--  .             USUBJID                       .            
#>  ID      ---  .             ID                            .            
#>  EVID    ---  .             event ID                      _yspec_lookup
#>  MDV     ---  .             MDV                           _yspec_lookup
#>  SEQ     -d-  .             SEQ                           .            
#>  AMT     ---  mg            amount                        .            
#>  II      ---  hours         II                            .            
#>  CMT     ---  .             Compartment                   .            
#>  TAFD    ---  hours         TAFD                          .            
#>  WT      ---  kg            Weight                        .            
#>  EGFR    ---  ml/min/1.73m2 eGFR                          _yspec_lookup
#>  STUDY   cd-  .             Study                         .            
#>  SEX     -d-  .             SEX                           _yspec_lookup
#>  BQL     -d-  .             below limit of quantification .            
#>  DV      ---  ng/mL         DV                            .            
#>  DV2     ---  .             DV2                           .            
#>  HAIR    cd-  .             HAIR                          .            
#>  CLCR    ---  ml/min        creatinine clearance          .            
#>  ALB     ---  g/dL          adjusted albumin              _yspec_lookup
fda_table(spec)
#>  [1] "\\begin{longtable}{|>{\\raggedright\\arraybackslash}p{0.75in}|>{\\raggedright\\arraybackslash}p{2.1in}|>{\\raggedright\\arraybackslash}p{0.6in}|>{\\raggedright\\arraybackslash}p{2.2in}|}"                                                                 
#>  [2] "  \\hline"                                                                                                                                                                                                                                                  
#>  [3] "VARIABLE & LABEL & TYPE & CODES \\\\ "                                                                                                                                                                                                                      
#>  [4] "  \\hline"                                                                                                                                                                                                                                                  
#>  [5] "\\endhead"                                                                                                                                                                                                                                                  
#>  [6] "\\hline"                                                                                                                                                                                                                                                    
#>  [7] "\\multicolumn{4}{l}{\\footnotesize Continued on next page}"                                                                                                                                                                                                 
#>  [8] "\\endfoot"                                                                                                                                                                                                                                                  
#>  [9] "\\endlastfoot"                                                                                                                                                                                                                                              
#> [10] " \\hline"                                                                                                                                                                                                                                                   
#> [11] "C & commented rows & character & values: ., C \\\\ "                                                                                                                                                                                                        
#> [12] "   \\hline"                                                                                                                                                                                                                                                 
#> [13] "USUBJID & universal subject identifier & character &  \\\\ "                                                                                                                                                                                                
#> [14] "   \\hline"                                                                                                                                                                                                                                                 
#> [15] "ID & NONMEM ID number & numeric &  \\\\ "                                                                                                                                                                                                                   
#> [16] "   \\hline"                                                                                                                                                                                                                                                 
#> [17] "EVID & event ID indicator & numeric &  \\\\ "                                                                                                                                                                                                               
#> [18] "   \\hline"                                                                                                                                                                                                                                                 
#> [19] "MDV & missing DV indicator & numeric &  \\\\ "                                                                                                                                                                                                              
#> [20] "   \\hline"                                                                                                                                                                                                                                                 
#> [21] "SEQ & record type indicators & numeric & 1 = concentration, 2 = response, 3 = other \\\\ "                                                                                                                                                                  
#> [22] "   \\hline"                                                                                                                                                                                                                                                 
#> [23] "AMT & dose amount (unit: mg) & numeric &  \\\\ "                                                                                                                                                                                                            
#> [24] "   \\hline"                                                                                                                                                                                                                                                 
#> [25] "II & interdose interval (unit: hours) & numeric &  \\\\ "                                                                                                                                                                                                   
#> [26] "   \\hline"                                                                                                                                                                                                                                                 
#> [27] "CMT & compartment number & numeric & values: 1, 2, 3, 4 \\\\ "                                                                                                                                                                                              
#> [28] "   \\hline"                                                                                                                                                                                                                                                 
#> [29] "TAFD & time after first dose (unit: hours) & numeric &  \\\\ "                                                                                                                                                                                              
#> [30] "   \\hline"                                                                                                                                                                                                                                                 
#> [31] "WT & baseline weight (unit: kg) & numeric &  \\\\ "                                                                                                                                                                                                         
#> [32] "   \\hline"                                                                                                                                                                                                                                                 
#> [33] "EGFR & estimated glomerular filtration rate (unit: ml/min/1.73m2) & numeric &  \\\\ "                                                                                                                                                                       
#> [34] "   \\hline"                                                                                                                                                                                                                                                 
#> [35] "STUDY & study identifier & character & 100 = The first phase 1 study, 202 = The first phase 2 study conducted only at sites 1, 2, 3 when the formulation was still oral liquid only, 303 = The renal impairment study, 203 = The second phase 2 study \\\\ "
#> [36] "   \\hline"                                                                                                                                                                                                                                                 
#> [37] "SEX & SEX & numeric & 0 = male, 1 = female \\\\ "                                                                                                                                                                                                           
#> [38] "   \\hline"                                                                                                                                                                                                                                                 
#> [39] "BQL & DV value is below the quantitation limit & numeric & 0, 1 \\\\ "                                                                                                                                                                                      
#> [40] "   \\hline"                                                                                                                                                                                                                                                 
#> [41] "DV & dependent variable (unit: ng/mL) & numeric &  \\\\ "                                                                                                                                                                                                   
#> [42] "   \\hline"                                                                                                                                                                                                                                                 
#> [43] "DV2 & second dependent variable & numeric &  \\\\ "                                                                                                                                                                                                         
#> [44] "   \\hline"                                                                                                                                                                                                                                                 
#> [45] "HAIR & patient hair color & character & 0 = brown, 1 = blonde, 2 = black \\\\ "                                                                                                                                                                             
#> [46] "   \\hline"                                                                                                                                                                                                                                                 
#> [47] "CLCR & creatinine clearance (unit: ml/min) & numeric &  \\\\ "                                                                                                                                                                                              
#> [48] "   \\hline"                                                                                                                                                                                                                                                 
#> [49] "ALB & adjusted albumin (unit: g/dL) & numeric &  \\\\ "                                                                                                                                                                                                     
#> [50] "  \\hline"                                                                                                                                                                                                                                                  
#> [51] "\\end{longtable}"