Write columns and column information for NONMEM $INPUT block

Column names are written to the console one-per-line on the left hand side of a NONMEM comment character (;) and a description of the column is written on the right hand side. The description includes the short name and values with decodes for non-character columns.

nm_input(spec, width = 60, cat = TRUE)

Arguments

spec
a yspec object

width
number of characters

cat
if TRUE, the text is sent to the console with cat()

Value

A character vector of text forming the $INPUT block (including the $INPUT part).

Examples

spec <- ys_help$spec()
nm_input(spec)

#> $INPUT
#> C         ; comment character
#> NUM       ; record number
#> ID        ; subject identifier
#> SUBJ=DROP ; subject identifier
#> TIME      ; TIME
#> SEQ       ; SEQ 
#>           ; [0 = observation, 1 = dose]
#> CMT       ; compartment number
#> EVID      ; event ID 
#>           ; [values: 0, 1]
#> AMT       ; dose amount
#> DV        ; dependent variable
#> AGE       ; age
#> WT        ; weight
#> CRCL      ; CRCL
#> ALB       ; albumin
#> BMI       ; BMI
#> AAG       ; alpha-1-acid glycoprotein
#> SCR       ; serum creatinine
#> AST       ; aspartate aminotransferase
#> ALT       ; alanine aminotransferase
#> HT        ; height
#> CP        ; Child-Pugh score 
#>           ; [0 = normal, 1 = Pugh1, 2 = Pugh2, 3 = Pugh3]
#> TAFD      ; time after first dose
#> TAD       ; time after dose
#> LDOS      ; last dose amount
#> MDV       ; MDV 
#>           ; [values: 0, 1]
#> BLQ       ; below limit of quantification 
#>           ; [1 = above QL, 0 = below Q]
#> PHASE     ; study phase indicator 
#>           ; [values: 1]
#> STUDY     ; study number 
#>           ; [1 = SAD, 2 = MAD, 3 = Renal, 4 = Hepatic]
#> RF=DROP   ; renal function stage