Render a define.pdf document

render_spec is an alias to render_define. See details.

render_define(x, ...)

# S3 method for class 'yproj'
render_define(
  x,
  stem = "define_working",
  format = "x_table_2",
  output_format = "pdf_document",
  output_dir = getwd(),
  build_dir = definetemplate(),
  title = "Data Specification",
  author = "",
  toc = "yes",
  number_sections = "yes",
  rmd_template = NULL,
  date = as.character(Sys.Date()),
  sponsor = NULL,
  projectnumber = NULL,
  ...
)

# S3 method for class 'character'
render_define(x, ..., dots = list())

# S3 method for class 'yspec'
render_define(...)

render_spec(x, ...)

# S3 method for class 'character'
render_spec(x, stem = basename(x), ..., dots = list())

# S3 method for class 'yspec'
render_spec(x, stem = get_meta(x)[["name"]], ..., dots = list())

Arguments

x

a yproj object or project specification file name

...

passed to rmarkdown::render()

stem

used to name the output file

format

the name of a function that will generate code formatting the data specification information

output_format

passed to rmarkdown::render()

output_dir

passed to rmarkdown::render()

build_dir

directory where rmarkdown should build the document

title

used in yaml front matter

author

used in yaml front matter

toc

used in yaml front matter

number_sections

used in yaml front matter

rmd_template

full path to rmarkdown file to be used to template the data specification document

date

used in yaml front matter

sponsor

optional project sponsor

projectnumber

optional project number in ...

dots

passed to object converter

Details

stem should not include a file extension, just the file stem.

latex requirements

For all document types, the following latex packages are required:

array

longtable

booktabs

fontenc

Make sure these packages are installed and available when trying to render a document.

Examples

file <- ys_help$file()

file
#> [1] "/tmp/RtmpPmqM4a/temp_libpath2a9e468d087c/yspec/internal/analysis1.yml"

if (FALSE) { # \dontrun{
  render_define(file)
} # }