yspec help object

ys_help is an environment containing functions that can be called to provide information and examples for how yspec works. specification(), internal_db() and analysis1() will open up markdown or yaml files using rstudioapi::navigateToFile if that package is installed and file.show in case it is not.

# S3 method for class 'yhelp'
print(...)

ys_help

Format

An object of class yhelp of length 12.

Arguments

...

not used

Details

spec() generates the analysis1 yspec object

file() the file name and path of an example yspec file

data() is a function that generates a data set that can be checked against spec()

csv() is the file name (and path) to the data set generated by data()

proj() generates a project object that can be rendered into a document

yaml() shows the analysis1 yaml file

ref() shows the syntax reference for yaml file

db() shows the internal column lookup data base

example() shows and Rmd file showing yspec workflow

export(path) export internal assets that can be used for exploration in the local filesystem; you must supply the location (path) to where the content is to be exported; there is also an overwrite argument that will allow overwriting an existing folder (use with care!)

Examples

spec <- ys_help$spec()

file <- ys_help$file()

data <- ys_help$data()  

csv  <- ys_help$csv()

if (FALSE) { # \dontrun{
proj <- ys_help$proj()

ref <- ys_help$ref()

db <- ys_help$db()

ex <- ys_help$example()

path <- ys_help$export("expath")

} # }