Generate content for FDA define document

Generate content for FDA define document

fda_define(file, title = "Datasets", ext = ".xpt", loc = ".", ...)

Arguments

file
full path to define yaml file

title
used in yaml front matter

ext
data set file extension to include; this should only be changed from default value of ".xpt" for testing purposes

loc
location to use for data set files; this should only be changed from default value of "." for testing purposes

...
arguments passed to rendering functions; see details

Value

A character vector of in markdown format. Wrap fda_define in writeLines and render asis in an Rmarkdown document.

See also

Examples

proj <- file_proj_ex()

cat(readLines(proj), sep = "\n")

#> # Generated by yspec::as_proj_spec; do not edit by hand
#> YPROJ__:
#>   sponsor: ABC-Pharma
#>   projectnumber: ABC101104F
#>   yml_file: /Users/kyleb/ghe/software/yspec/inst/spec/project.yml
#>   proj_file: /Users/kyleb/ghe/software/yspec/inst/spec/project.yml
#>   proj_path: /Users/kyleb/ghe/software/yspec/inst/spec
#>   path: /Users/kyleb/ghe/software/yspec/inst/spec
#>   class: yproj
#> DEM104101F_PK:
#>   name: DEM104101F_PK
#>   description: Population PK analysis data set
#>   spec_path: /Users/kyleb/ghe/software/yspec/inst/spec
#>   spec_file: /Users/kyleb/ghe/software/yspec/inst/spec/DEM104101F_PK.yml
#>   data_path: ../data/derived
#>   data_stem: DEM104101F_PK
#>   csv_file: ../data/derived/DEM104101F_PK.csv
#>   xpt_file: ../data/derived/DEM104101F_PK.xpt
#> DEM104101F_PKPD:
#>   name: DEM104101F_PKPD
#>   description: Population PKPD analysis data set
#>   spec_path: /Users/kyleb/ghe/software/yspec/inst/spec
#>   spec_file: /Users/kyleb/ghe/software/yspec/inst/spec/DEM104101F_PKPD.yml
#>   data_path: ../data/derived
#>   data_stem: DEM104101F_PKPD
#>   csv_file: ../data/derived/DEM104101F_PKPD.csv
#>   xpt_file: ../data/derived/DEM104101F_PKPD.xpt
#> DEM104101F_AE:
#>   name: DEM104101F_AE
#>   description: AE analysis data set
#>   spec_path: /Users/kyleb/ghe/software/yspec/inst/spec
#>   spec_file: /Users/kyleb/ghe/software/yspec/inst/spec/DEM104101F_AE.yml
#>   data_path: ../data/derived
#>   data_stem: DEM0104101F_AE_2
#>   csv_file: ../data/derived/DEM0104101F_AE_2.csv
#>   xpt_file: ../data/derived/DEM0104101F_AE_2.xpt
#> 

if (FALSE) {
  fda_define(proj)
}